(PC-27682)[PRO] refactor: final quality cleanup
