+++ pro/src/pages/Offerers/Offerer/VenueV1/fields/ReimbursementFields/ReimbursementFields.tsx
@@ -2,8 +2,8 @@ import React, { useCallback, useState } from 'react'
 
 import { GetOffererResponseModel, GetVenueResponseModel } from 'apiClient/v1'
 import FormLayout from 'components/FormLayout'
-import ReimbursementPoint from 'pages/VenueCreation/ReimbursementPoint/ReimbursementPoint'
 import useActiveFeature from 'hooks/useActiveFeature'
+import ReimbursementPoint from 'pages/VenueCreation/ReimbursementPoint/ReimbursementPoint'
 import { Banner } from 'ui-kit'
 
 import PricingPoint from '../PricingPoint'
