--- pro/src/pages/VenueCreation/ImageUploaderVenue/ImageUploaderVenue.module.scss
@@ -1,19 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-
-.explanatory-text {
-  margin-bottom: rem.torem(24px);
-}
-
-.image-venue-uploader-section {
-  &-image-container {
-    display: flex;
-  }
-
-  &-icon-container {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: space-evenly;
-    margin-left: rem.torem(32px);
-  }
-}
