+++ pro/src/screens/SignupJourneyForm/Offerer/__specs__/Offerer.spec.tsx
@@ -8,7 +8,7 @@ import { api } from 'apiClient/api'
 import { ApiError } from 'apiClient/v1'
 import { ApiRequestOptions } from 'apiClient/v1/core/ApiRequestOptions'
 import { ApiResult } from 'apiClient/v1/core/ApiResult'
-import { DEFAULT_ADDRESS_FORM_VALUES } from 'components/Address'
+import { DEFAULT_ADDRESS_FORM_VALUES } from 'components/Address/constants'
 import Notification from 'components/Notification/Notification'
 import {
   SignupJourneyContext,
